Add params to HDF5ComparisonTool
This adds a tolerance parameter to check values agree up to a tolerance
and an abort_limit parameter which is the maximum number of
disagreements between values on a level before aborting.
The return code has also been removed as MPI implementations sometimes
complain and the output to stdout is sufficient.